feat(handleCancel): export handleConfirmCancel function
